refactor(cb2-2829): module rename to better reflect usage
